- moved AppInjector to package com.erm.artists.di
